Some basic and some more advanced canvas.
Plus this is Object oriented, meaning I am using the new Javascript version (idk how new) which means I can make classes with methods and variables, in a seperate file, which is nice.

This has some different js scripts. some more advanced than others. The advanced one is one where 1000 circle objects are created, and these objects is moving around in different speeds. These objects are updated 50 times a second, meaning we are cleaning the canvas and drawing their new x and y values.
If the mouse is in a radius of 50, then the circles will expand to their max radius (which also is random).