Added some ugly animations when walking right.
I don't think I will be working much more on this, as canvas games are outdated and is kind of a waste of time to work on this much.

The point was to make a multiplayer platformer where you could shoot each other, but when it is this hard making it single player, I don't think I am ready making it multiplayer.

But I learned a lot about javascript classes and object oriented programming.